rename proto_parser package
