+++ adapter/parser/parser.go
@@ -1,13 +1,13 @@
 package parser
 
 import (
-	"github.com/ktr0731/evans/adapter/internal/proto_parser"
+	"github.com/ktr0731/evans/adapter/internal/protoparser"
 	"github.com/ktr0731/evans/adapter/protobuf"
 	"github.com/ktr0731/evans/entity"
 )
 
 func ParseFile(filename []string, paths []string) ([]*entity.Package, error) {
-	set, err := proto_parser.ParseFile(filename, paths)
+	set, err := protoparser.ParseFile(filename, paths)
 	if err != nil {
 		return nil, err
 	}
